requirements_sms.txt
requests==2.32.3
python-dotenv==1.0.1
twilio==9.0.0

flask==3.0.0

